Disabled Hyprland/Fabric
The fabric widgets I wanted to use are based on Fabric1, which is now
deprecated. I'll probably just wait until it's rebuilt for fabric2,
because it's not worth it as is.

Note for myself: I'm manually building fabric-cli from the github repo,
and installing fabric from the github flake. Why these two aren't
packaged together is beyond me.